requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
bidict==0.23.1
bracex==2.5.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
frozenlist==1.5.0
ghp-import==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.8.0
lazydocs==0.4.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-awesome-pages-plugin==2.9.3
mkdocs-get-deps==0.2.0
multidict==6.1.0
natsort==8.4.0
openai==1.57.0
packaging==24.2
pathspec==0.12.1
payments-py==0.5.3
pinatapy-vourhey==0.2.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
pyyaml_env_tag==0.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==6.0.0
wcmatch==10.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
